Update UI Screens in Compose
